% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lineChartData.R
\name{lineChartData}
\alias{lineChartData}
\title{Line chart data}
\usage{
lineChartData(x, y, key, color, area = FALSE)
}
\arguments{
\item{x}{values on the x-axis, numeric or date type}

\item{y}{values on the y-axis}

\item{key}{string, the title of the line chart}

\item{color}{line chart color}

\item{area}{Boolean, whether to turn the line chart into a filled area chart}
}
\value{
A list, for usage in \code{\link{lineChart}}.
}
\description{
Make line chart data.
}
