% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnvd3.R
\name{multiBarChart}
\alias{multiBarChart}
\title{xxx}
\usage{
multiBarChart(
  data,
  formula,
  by,
  palette = "viridis",
  xAxisTitle = NULL,
  yAxisTitle = NULL,
  margins = list(b = 100, l = 70),
  duration = 1300,
  rotateLabels = 0,
  groupSpacing = 0.1,
  xAxisLabelDistance = 35,
  yAxisLabelDistance = -5,
  yAxisShowMaxMin = FALSE,
  yAxisTickFormat = "d",
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{data}{dataframe used for the chart}

\item{formula}{a two-sided formula like \code{y ~ x}, where \code{"x"} and
\code{"y"} are two column names of \code{data}}

\item{by}{string, the "by" variable; must be a column name of \code{data}}

\item{palette}{the name of a viridis color palette, e.g. \cod{"viridis"},
\code{"cividis"} or \code{"turbo"};
see \code{\link[viridisLite]{viridis}}}

\item{xAxisTitle}{a title for the x-axis; if \code{NULL}, the title is
taken from the \code{formula} argument}

\item{yAxisTitle}{a title for the y-axis; if \code{NULL}, the title is
taken from the \code{formula} argument}

\item{margins}{a named list defining the margins, with names \code{"t"},
\code{"r"}, \code{"b"} and \code{"l"}, for "top", "right", "bottom"
and "left" respectively}

\item{duration}{duration of the transition, a number of milliseconds}

\item{rotateLabels}{a number, the angle of rotation of the labels of the
x-axis (in degrees)}

\item{yAxisShowMaxMin}{Boolean, whether to show the min and the max on
the y-axis}

\item{elementId}{}
}
\value{
A htmlwidget displaying a grouped/stacked bar chart.
}
\description{
xxx
}
\examples{
library(Rnvd3)
dat <- reshape2::melt(
  apply(HairEyeColor, c(1, 2), sum), value.name = "Count"
)
multiBarChart(dat, Count ~ Eye, by = "Hair")
}
