% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnvd3.R
\name{lineChart}
\alias{lineChart}
\title{Title}
\usage{
lineChart(
  data,
  xAxisTitle = "x",
  yAxisTitle = "y",
  margins = list(l = 100),
  duration = 500,
  useInteractiveGuideline = TRUE,
  width = NULL,
  height = NULL,
  elementId = NULL
)
}
\arguments{
\item{elementId}{}
}
\value{

}
\description{
Title
}
